+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Support for Excel coordinates import
 - Support for DCS Scratchpad coordinates import
 - Support for Aerodromes excel coordinates import
+- Bingo command for DCS Scratchpad data file
 
 ### Changed
 - Using Python argparse to parse commandline arguments